fix graphqltypes when running in production mode
